requirements.txt
appdirs==1.4.4
asgiref==3.5.2
attrs==22.1.0
cattrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
exceptiongroup==1.0.0rc8
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
requests-cache==0.9.6
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
url-normalize==1.4.3
urllib3==1.26.12
